refactor: split auth response and json helpers
